feat: allow minting NUCs in `nillion` cli
