+++ tools/nillion/Cargo.toml
@@ -10,6 +10,7 @@ config = { version = "0.14", default-features = false, features = ["yaml"] }
 env_logger = "0.11.5"
 chrono = { version = "0.4", features = ["serde"] }
 hex = { version = "0.4", features = ["serde"] }
+humantime = "2.2"
 futures = "0.3.30"
 log = "0.4"
 nillion-nucs = { path = "../../libs/nucs" }
