+++ tools/nillion/src/lib.rs
@@ -5,7 +5,7 @@ use std::{fs::File, io::BufReader, path::Path};
 pub mod args;
 pub mod config;
 pub mod context;
-pub mod runner;
+pub mod handlers;
 pub mod serialize;
 pub(crate) mod wrappers;
